Move wrapper code into wrapper/ to make it compliant with inner Papillon directory structure
